First attempt at achat.py
achat.py is an async(-ish) version of chat.py that reads one key at a
time instead of one line at a time. This allows one side to send
messages to the other without blocking.

Note that this version is a bad, laggy typing experience. It could use a
lot of work.